Move deadlock unit test to a separate package
